% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkDeplete-package.R
\docType{package}
\name{LinkDeplete-package}
\alias{LinkDeplete}
\alias{LinkDeplete-package}
\title{LinkDeplete: Within-Season Depletion Model for Crustacean Stock Assessment}
\description{
Fits a Leslie depletion model linked across seasons via a random walk on recruitment using RTMB. Supports multiple CPUE series (depletion and biomass indices), time-varying catchability, and forward projection to target reference points. Includes diagnostic plotting and privacy-mode output for presentations.
}
\author{
\strong{Maintainer}: Simon de Lestang \email{simon.delestang@dpird.wa.gov.au}

Authors:
\itemize{
  \item Simon de Lestang \email{simon.delestang@dpird.wa.gov.au}
}

}
\keyword{internal}
